requirements.txt
Django==1.11.21
django-countries==5.3.3
Pillow==6.0.0
pytz==2019.1
